bot/requirements.txt
discord.py
mcrcon